Move AnalyzerClientMock to mocks.dart
