+++ app/test/frontend/mocks.dart
@@ -2,8 +2,12 @@
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
 
+import 'package:pub_dartlang_org/dartdoc/models.dart';
 import 'package:pub_dartlang_org/scorecard/backend.dart';
 import 'package:pub_dartlang_org/scorecard/models.dart';
+import 'package:pub_dartlang_org/shared/analyzer_service.dart';
+import 'package:pub_dartlang_org/shared/analyzer_client.dart';
+import 'package:pub_dartlang_org/shared/dartdoc_client.dart';
 
 class ScoreCardBackendMock implements ScoreCardBackend {
   @override
@@ -34,10 +38,10 @@ class ScoreCardBackendMock implements ScoreCardBackend {
   Future<bool> shouldUpdateReport(
       String packageName, String packageVersion, String reportType,
       {bool includeDiscontinued = false,
-        bool includeObsolete = false,
-        Duration successThreshold = const Duration(days: 30),
-        Duration failureThreshold = const Duration(days: 1),
-        DateTime updatedAfter}) {
+      bool includeObsolete = false,
+      Duration successThreshold = const Duration(days: 30),
+      Duration failureThreshold = const Duration(days: 1),
+      DateTime updatedAfter}) {
     throw new UnimplementedError();
   }
 
@@ -50,5 +54,51 @@ class ScoreCardBackendMock implements ScoreCardBackend {
   @override
   Future updateScoreCard(String packageName, String packageVersion) {
     throw new UnimplementedError();
+  }
+}
+
+class AnalyzerClientMock implements AnalyzerClient {
+  AnalysisView mockAnalysisView;
+
+  @override
+  Future close() async => null;
+
+  @override
+  Future<AnalysisView> getAnalysisView(AnalysisKey key) async =>
+      mockAnalysisView;
+
+  @override
+  Future<List<AnalysisView>> getAnalysisViews(Iterable<AnalysisKey> keys) =>
+      Future.wait(keys.map(getAnalysisView));
+
+  @override
+  Future triggerAnalysis(
+      String package, String version, Set<String> dependentPackages) async {}
+}
+
+class DartdocClientMock implements DartdocClient {
+  @override
+  Future<DartdocEntry> getEntry(String package, String version) async {
+    return null;
+  }
+
+  @override
+  Future<List<DartdocEntry>> getEntries(
+      String package, List<String> versions) async {
+    return versions.map((s) => null).toList();
+  }
+
+  @override
+  Future triggerDartdoc(
+      String package, String version, Set<String> dependentPackages) async {}
+
+  @override
+  Future close() async {}
+
+  @override
+  Future<String> getTextContent(
+      String package, String version, String relativePath,
+      {Duration timeout}) async {
+    return null;
   }
 }
